requirements.txt
streamlit
google-genai
pydantic
pandas
openpyxl